transport: Reviewed communicator rekey test. Fixes #6828
The TCP communicator already expires keys according to config, but it
was not possible to configure the maximum rekey bytes.
Note that we also still do not have test that tests this specifically.
Conversely, the UDP communicator test tests the expiration after the
number of bytes, but the option for the expiration is not used still.